% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.StratifiedXValS3.R
\name{predict.StratifiedXValS3}
\alias{predict.StratifiedXValS3}
\title{Obtain a cross validation fold from data based on an S\code{stratifiedXValS3} object.}
\usage{
\method{predict}{StratifiedXValS3}(obj, fold, data)
}
\arguments{
\item{obj}{An object of type \code{StratifiedXValS3}.}

\item{fold}{The index of the fold you which to retrieve.}

\item{data}{The original data that the folds were computed on.}
}
\value{
A list containing two dataframes, \code{Train} and \code{Test}.
}
\description{
Returns a list of Training and Testing data based on folds which have been previously computed by \code{stratifySample()}.
}
\examples{
predict()

}
