% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_sample.R
\name{stratifySample}
\alias{stratifySample}
\title{Performs stratified cross validation sampling of a data frame.}
\usage{
stratifySample(data, targetVariable, nfolds = 5, uniqueID = NULL,
  keepData = FALSE, positiveResponse = 1, seed = NULL)
}
\arguments{
\item{data}{A \code{dataframe} which contains the base data to be split.}

\item{targetVariable}{A \code{string} giving the name of the binary variable to stratify by.}

\item{nfolds}{An integer giving the number of cross validation folds.}

\item{uniqueID}{A \code{string} giving the name of a unique row identifier. Note, if there is no identifier, the function
will filter the dataset by index.}

\item{keepData}{A Boolean indicating whether or not to keep the entire dataset along with extra columns indicating the folds.}

\item{positiveResponse}{The value of the target variable which indicates a positive response. Must be 1 or 0.}

\item{seed}{The seed value to use to reproduce results.}
}
\value{
An object of the class \code{StratifiedXValS3}.
}
\description{
Generates cross validation folds in a stratified way.
}
\examples{
stratifySample()

}
